Refactored set_led_circle_dir_dist function to helper functions.
