+++ Project/main/src/helperFunctions.c
@@ -3,6 +3,18 @@
 LOG_MODULE_REGISTER(helperFunctions);
 K_TIMER_DEFINE(timer, NULL, NULL);
 
+// LED circle direction to location related definitions
+#define ANGLE_BUFFER_SIZE 20 // Number of angles to average for the circle direction
+#define DIST_MAX_WIDTH 20	// Distance at which the circle has minimum width
+#define DIST_MIN_WIDTH 200 	//Distance at which the circle has maximum width
+#define DIST_RANGE (DIST_MIN_WIDTH - DIST_MAX_WIDTH)
+#define LEDS_MIN_WIDTH 0	// Amount of LEDS at DIST_MIN_WIDTH
+#define LEDS_MAX_WIDTH 48	// Amount of LEDS at DIST_MAX_WIDTH
+#define LEDS_RANGE_WIDTH (LEDS_MAX_WIDTH - LEDS_MIN_WIDTH)
+#define MAX_DIR 359
+#define NR_OF_LEDS 64
+
+
 //Convert btnmatrix input to btnmatrix output format
 void btnmatrix_in_to_out(uint8_t* btnmatrix_input, uint8_t* btnmatrix_output){
 	for (uint8_t i = 0; i < 16; i++) {
@@ -223,4 +235,101 @@ void set_btnmatrix_led(uint8_t position){
 	data_button_matrix[row] |= (1 << column);
 
 	btnmatrix_outSetMutexValue(data_button_matrix);
+}
+
+/**
+ * @brief Convert degrees to radians.
+ * @param degree The angle in degrees.
+ * @return The angle in radians.
+ */
+int get_led_width(int distance) {
+	if (distance > DIST_MIN_WIDTH) { return LEDS_MIN_WIDTH; }
+	if (distance < DIST_MAX_WIDTH) { return LEDS_MAX_WIDTH; }
+	distance -= DIST_MAX_WIDTH;
+	double distFraction = (double)distance / DIST_RANGE;
+	distFraction = 1 - distFraction;
+	return LEDS_RANGE_WIDTH*distFraction + LEDS_MIN_WIDTH;
+}
+
+/**
+ * @brief Convert degrees to radians.
+ * @param degree The angle in degrees.
+ * @return The angle in radians.
+ */
+int circleMovingAvg(int newValue) {
+	double rad = toRadians(newValue);
+	double sinVal = sin(rad);
+	double cosVal = cos(rad);
+
+	static double angleSinBuffer[ANGLE_BUFFER_SIZE];
+	static double angleCosBuffer[ANGLE_BUFFER_SIZE];
+
+	for (int i = ANGLE_BUFFER_SIZE-1; i > 0; i--) {	// Move every entry in the buffer up by one index
+		angleSinBuffer[i] = angleSinBuffer[i-1];
+		angleCosBuffer[i] = angleCosBuffer[i-1];
+	}
+	angleSinBuffer[0] = sinVal;						// Insert the new value
+	angleCosBuffer[0] = cosVal;
+
+	double sinSum = 0;
+	double cosSum = 0;
+	for (int i = 0; i < ANGLE_BUFFER_SIZE; i++) {
+		sinSum += angleSinBuffer[i];
+		cosSum += angleCosBuffer[i];
+	}
+	double sinAvg = sinSum / ANGLE_BUFFER_SIZE;
+	double cosAvg = cosSum / ANGLE_BUFFER_SIZE;
+	return toDegrees(atan2(sinAvg,cosAvg));
+}
+
+/**
+ * @brief Function to set the led circle direction and distance for showing where the user should go next.
+ * @param dir The direction in degrees (0-359) where the LEDs should be centered.
+ * @param distance The distance in meters to the target location, which will determine the width of the LED circle.
+ */
+void set_led_circle_dir_dist(int dir, int distance)
+{
+	int width = get_led_width(distance); // Convert distance to width in pixels
+
+	float centerFloat = (float)dir / MAX_DIR;
+	int centerPixel = round(centerFloat * NR_OF_LEDS);
+
+	int leftBound = centerPixel - (width / 2);
+	if (leftBound < 0 ) {
+		leftBound = NR_OF_LEDS + leftBound;
+	}
+	int rightBound = centerPixel + (width / 2);
+	if (rightBound > (NR_OF_LEDS - 1)) {
+		rightBound = rightBound - NR_OF_LEDS;
+	}
+	bool overlap = false;
+	if (leftBound > rightBound) {
+		overlap = true;
+	}
+
+	const int nrBytes = 8;
+	const int nrBits = 8;
+	uint8_t outputValues[8] = {0,0,0,0,0,0,0,0};	// Initialize to zero (all LEDs off)
+	for (int byteCount = 0; byteCount < nrBytes; byteCount++) {
+		for (int bitCount = 0; bitCount < nrBits; bitCount++) {
+			int bitIndex = 8 * byteCount + bitCount;
+			//bool enableLed = (!overlap && (bitIndex >= leftBound) && (bitIndex >= rightBound)) || (overlap && (bitIndex >= leftBound || bitIndex <= rightBound));
+			bool enableLed = false;
+			if (!overlap) {
+				if (bitIndex >= leftBound && bitIndex <= rightBound) {
+					enableLed = true;
+				}
+			} else {
+				if (bitIndex >= leftBound || bitIndex <= rightBound) {
+					enableLed = true;
+				}
+			}
+			//LOG_WRN("Byte %d bit %d (bitindex %d): %d\n", byteCount, bitCount, bitIndex, enableLed);
+			if (enableLed) {
+				outputValues[byteCount] |= (1 << (7-bitCount));
+			}
+
+		}
+	}
+	ledcircleSetMutexValue(outputValues);
 }
